Add difference between StationToView and Station and Add blank fragment for APP INFO

+++ android/app/src/main/java/com/ismin/csproject/Station.kt
@@ -1,15 +1,15 @@
 package com.ismin.csproject
 
 data class Station(
-    val ylatitude : Float,
-    val xlatitude : Float,
-    val type_prise : String,
-    val id_station : String,
+    val ylatitude: Double,
+    val xlatitude: Double,
+    val type_prise: String,
+    val id_station: String,
 
     val accessibilite: String,
-    val region : String,
-    val ad_station : String,
-    val puiss_max: Float,
+    val region: String,
+    val ad_station: String,
+    val puiss_max: Double,
     val acces_recharge: String,
-    val nbre_pdc: Float
+    val nbre_pdc: Double
 )
